# IPC_HLE: Don't open/close files for every single file operation.