--- a/Source/Core/Core/IPC_HLE/WII_IPC_HLE.h
+++ b/Source/Core/Core/IPC_HLE/WII_IPC_HLE.h
@@ -4,6 +4,7 @@
 
 #pragma once
 
+#include <memory>
 #include <string>
 
 #include "Common/CommonTypes.h"
@@ -63,10 +64,10 @@ void ES_DIVerify(u8 *_pTMD, u32 _sz);
 void SDIO_EventNotify();
 
 
-IWII_IPC_HLE_Device* CreateFileIO(u32 _DeviceID, const std::string& _rDeviceName);
+std::shared_ptr<IWII_IPC_HLE_Device> CreateFileIO(u32 _DeviceID, const std::string& _rDeviceName);
 
-IWII_IPC_HLE_Device* GetDeviceByName(const std::string& _rDeviceName);
-IWII_IPC_HLE_Device* AccessDeviceByID(u32 _ID);
+std::shared_ptr<IWII_IPC_HLE_Device> GetDeviceByName(const std::string& _rDeviceName);
+std::shared_ptr<IWII_IPC_HLE_Device> AccessDeviceByID(u32 _ID);
 int getFreeDeviceId();
 
 // Update