--- a/Source/Core/Core/IPC_HLE/WII_IPC_HLE_Device_FileIO.h
+++ b/Source/Core/Core/IPC_HLE/WII_IPC_HLE_Device_FileIO.h
@@ -28,7 +28,7 @@ public:
 	IPCCommandResult IOCtl(u32 _CommandAddress) override;
 	void DoState(PointerWrap &p) override;
 
-	File::IOFile OpenFile();
+	void OpenFile();
 
 private:
 	enum
@@ -75,4 +75,5 @@ private:
 	u32 m_SeekPos;
 
 	std::string m_filepath;
+	File::IOFile m_file;
 };