--- a/Source/Core/Core/IPC_HLE/WII_IPC_HLE_Device_FileIO.h
+++ b/Source/Core/Core/IPC_HLE/WII_IPC_HLE_Device_FileIO.h
@@ -75,5 +75,5 @@ private:
 	u32 m_SeekPos;
 
 	std::string m_filepath;
-	File::IOFile m_file;
+	std::shared_ptr<File::IOFile> m_file;
 };