This *weird* behaviour was originally introduced to fix a issue with PokéPark Wii (and other games) but it causes some performance issues, especially if you have a virus scanning intercepting all open/close calls.

Commit one simplifies the lifecycle of IPC_HLE devices using shared pointers, fixing a memory leak and making it harder to accidentally leak memory/file handles in the savestate code.

Commit two reverts to the original behaviour, breaking PokéPark but decreasing the Pokémon Snap initial boot time from 3+ min to a few seconds on a default windows 8 install.

Commit three implements a new fix for PokéPark et al, making sure dolphin only opens files once, even if the game opens the same file multiple times.

Fixes [Issue 8702](https://code.google.com/p/dolphin-emu/issues/detail?id=8702), Re-fixes [Issue 2917](https://code.google.com/p/dolphin-emu/issues/detail?id=2917) and [Issue 5232](https://code.google.com/p/dolphin-emu/issues/detail?id=5232)