@phire 2015-06-23
@Armada651 I'm not sure if you want to merge it into stable or not.

Ultimately it's a performance fix and the code it touches it somewhat risky.