@Armada651 2015-06-23
I don't think we should merge more performance fixes, let's concentrate on regressions. Instead of merging performance fixes after the feature freeze we should just do more stable releases.