@BhaaLseN 2015-06-23
Source-wise LGTM (not tested).

And I agree with with Armada, lets try to get stable out the door (with regression fixes as short-term necessary) and resume our release policy discussion to find a path that allows us to push out stable releases in a shorter time frame.