@i30817 2015-06-23
I wonder if this will affect this: https://code.google.com/p/dolphin-emu/issues/detail?id=8503&can=4
the dev said the log indicated eternal writes to memory card.