@mbc07 2015-06-23
Hey guys, did a quick test with Wii Menu and Super Smash Bros Brawl and this PR also fixes [issue 6829](https://code.google.com/p/dolphin-emu/issues/detail?id=6829)! YAY