@phire 2015-06-24
@Armada651 I'm going to have to argue that this doesn't count as fixing a performance regression.

A performance regression would be dropping from 250fps to 190fps or even 60fps to 20fps. 

This is increasing various 0.1 to 3.0 second lockups of the dolphin emulation to several minutes, giving the appearance that dolphin has crashed. If the user attempts to stop emulation or kill dolphin due to this apparent crash, their save file gets corrupted, requiring manual deletion before the game will work again.

I'm not saying this should be merged into the 5.0 stable, just that "only a performance regression" is not the best argument.