@phire 2015-06-24
The 4th/last commit has been redone with templates, to address some concerns. After that is reviewed this PR should be ready for merge.