@i30817 2015-07-03
This doesn't fix issue 8503 on google code. I added a save file to that issue that can reproduce it there.