@phire 2015-07-04
Fixed typo and removed non-acsii characters from a comment, which were causing issues in some IDEs